+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-client/src/test/java/org/allGraphQLCases/IntrospectionIT.java
@@ -41,10 +41,10 @@ import graphql.introspection.IntrospectionQuery;
 @Execution(ExecutionMode.CONCURRENT)
 public class IntrospectionIT {
 
-	static String[] AllFieldCases_FIELDS = { "id", "name", "forname", "age", "aFloat", "date", "dateTime", "dates",
-			"nbComments", "comments", "booleans", "aliases", "planets", "friends", "matrix", "oneWithIdSubType",
-			"listWithIdSubTypes", "oneWithoutIdSubType", "listWithoutIdSubTypes", "issue65", "issue66",
-			"extendedField" };
+	static String[] AllFieldCases_FIELDS = { "id", "name", "forname", "break", "age", "aFloat", "date", "dateTime",
+			"dates", "nbComments", "comments", "booleans", "aliases", "planets", "friends", "matrix",
+			"oneWithIdSubType", "listWithIdSubTypes", "oneWithoutIdSubType", "listWithoutIdSubTypes", "issue65",
+			"issue66", "extendedField" };
 	static List<String> AllFieldCases_FIELDNAMES = Arrays.asList(AllFieldCases_FIELDS);
 
 	@Autowired